Authorization using pundit

+++ app/controllers/favorite_stocks_controller.rb
@@ -1,8 +1,10 @@
 class FavoriteStocksController < ApplicationController
   before_action :authenticate_user!, only: %i[create destroy]
+  after_action :verify_authorized
 
   def create
     @stock_id = params[:stock_id]
+    authorize FavoriteStock
     current_user.favorite_stocks.create(stock_id: @stock_id)
     respond_to do |format|
       format.js
@@ -12,6 +14,7 @@ class FavoriteStocksController < ApplicationController
   def destroy
     @stock_id = params[:id]
     favorite_stock = current_user.favorite_stocks.find_by(stock_id: @stock_id)
+    authorize favorite_stock
     favorite_stock.destroy
     respond_to do |format|
       format.js
